+++ consensus/common_ifaces.go
@@ -1,0 +1,10 @@
+package consensus
+
+import (
+	"github.com/kaiachain/kaia/blockchain/types"
+)
+
+type Verifier interface {
+	VerifyHeader(chain ChainReader, header *types.Header) error
+	VerifySeals(chain ChainReader, header *types.Header) error
+}
