+++ consensus/consensus.go
@@ -76,6 +76,8 @@ type ChainReader interface {
 //
 //go:generate mockgen -destination=./mocks/engine_mock.go -package=mocks github.com/kaiachain/kaia/consensus Engine
 type Engine interface {
+	Verifier
+
 	// Author retrieves the Kaia address of the account that minted the given
 	// block.
 	Author(header *types.Header) (common.Address, error)
@@ -85,21 +87,6 @@ type Engine interface {
 
 	// PreprocessHeaderVerification prepares header verification for heavy computation before synchronous header verification such as ecrecover.
 	PreprocessHeaderVerification(headers []*types.Header) (chan<- struct{}, <-chan error)
-
-	// VerifyHeader checks whether a header conforms to the consensus rules of a
-	// given engine. Verifying the seal may be done optionally here, or explicitly
-	// via the VerifySeal method.
-	VerifyHeader(chain ChainReader, header *types.Header, seal bool) error
-
-	// VerifyHeaders is similar to VerifyHeader, but verifies a batch of headers
-	// concurrently. The method returns a quit channel to abort the operations and
-	// a results channel to retrieve the async verifications (the order is that of
-	// the input slice).
-	VerifyHeaders(chain ChainReader, headers []*types.Header, seals []bool) (chan<- struct{}, <-chan error)
-
-	// VerifySeal checks whether the crypto seal on a header is valid according to
-	// the consensus rules of the given engine.
-	VerifySeal(chain ChainReader, header *types.Header) error
 
 	// Prepare initializes the consensus fields of a block header according to the
 	// rules of a particular engine. The changes are executed inline.
