+++ consensus/istanbul/backend/verifier.go
@@ -1,0 +1,224 @@
+package backend
+
+import (
+	"math/big"
+
+	"github.com/kaiachain/kaia/blockchain/types"
+	"github.com/kaiachain/kaia/consensus"
+	istanbulCore "github.com/kaiachain/kaia/consensus/istanbul/core"
+	"github.com/kaiachain/kaia/consensus/misc"
+	"github.com/kaiachain/kaia/consensus/misc/eip4844"
+)
+
+var _ consensus.Verifier = &backend{}
+
+func (sb *backend) VerifyHeader(chain consensus.ChainReader, header *types.Header) error {
+	var parent []*types.Header
+	if header.Number.Sign() == 0 {
+		// If current block is genesis, the parent is also genesis
+		parent = append(parent, chain.GetHeaderByNumber(0))
+	} else {
+		parent = append(parent, chain.GetHeader(header.ParentHash, header.Number.Uint64()-1))
+	}
+	return sb.verifyHeader(chain, header, parent)
+}
+
+func (sb *backend) VerifySeals(chain consensus.ChainReader, header *types.Header) error {
+	// get parent header and ensure the signer is in parent's validator set
+	number := header.Number.Uint64()
+	if number == 0 {
+		return errUnknownBlock
+	}
+
+	// ensure that the blockscore equals to defaultBlockScore
+	if header.BlockScore.Cmp(defaultBlockScore) != 0 {
+		return errInvalidBlockScore
+	}
+	return sb.verifySigner(chain, header, nil)
+}
+
+// verifyHeader checks whether a header conforms to the consensus rules.The
+// caller may optionally pass in a batch of parents (ascending order) to avoid
+// looking those up from the database. This is useful for concurrently verifying
+// a batch of new headers.
+func (sb *backend) verifyHeader(chain consensus.ChainReader, header *types.Header, parents []*types.Header) error {
+	if header.Number == nil {
+		return errUnknownBlock
+	}
+
+	// Header verify before/after magma fork
+	if chain.Config().IsMagmaForkEnabled(header.Number) {
+		if len(parents) > 0 {
+			// the kip71Config used when creating the block number is a previous block config.
+			blockNum := header.Number.Uint64()
+			pset := sb.govModule.GetParamSet(blockNum)
+			kip71 := pset.ToKip71Config()
+			if err := misc.VerifyMagmaHeader(parents[len(parents)-1], header, kip71); err != nil {
+				return err
+			}
+		}
+		// For Magma fork, BaseFee is allowed even without parents (first header)
+	} else if header.BaseFee != nil {
+		return consensus.ErrInvalidBaseFee
+	}
+
+	// Don't waste time checking blocks from the future
+	if header.Time.Cmp(big.NewInt(now().Add(allowedFutureBlockTime).Unix())) > 0 {
+		return consensus.ErrFutureBlock
+	}
+
+	// Ensure that the extra data format is satisfied
+	if _, err := types.ExtractIstanbulExtra(header); err != nil {
+		return errInvalidExtraDataFormat
+	}
+	// Ensure that the block's blockscore is meaningful (may not be correct at this point)
+	if header.BlockScore == nil || header.BlockScore.Cmp(defaultBlockScore) != 0 {
+		return errInvalidBlockScore
+	}
+
+	// TODO-kaiax: further flatten the code inside; especially after most of the checks are moved to consensus modules
+	if err := sb.verifyCascadingFields(chain, header, parents); err != nil {
+		return err
+	}
+
+	for _, module := range sb.consensusModules {
+		if err := module.VerifyHeader(header); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// verifyCascadingFields verifies all the header fields that are not standalone,
+// rather depend on a batch of previous headers. The caller may optionally pass
+// in a batch of parents (ascending order) to avoid looking those up from the
+// database. This is useful for concurrently verifying a batch of new headers.
+func (sb *backend) verifyCascadingFields(chain consensus.ChainReader, header *types.Header, parents []*types.Header) error {
+	// The genesis block is the always valid dead-end
+	number := header.Number.Uint64()
+	if number == 0 {
+		return nil
+	}
+	// Ensure that the block's timestamp isn't too close to it's parent
+	var parent *types.Header
+	if len(parents) > 0 {
+		parent = parents[len(parents)-1]
+	} else {
+		parent = chain.GetHeader(header.ParentHash, number-1)
+	}
+	if parent == nil || parent.Number.Uint64() != number-1 || parent.Hash() != header.ParentHash {
+		return consensus.ErrUnknownAncestor
+	}
+	if parent.Time.Uint64()+sb.config.BlockPeriod > header.Time.Uint64() {
+		return errInvalidTimestamp
+	}
+	if err := sb.verifySigner(chain, header, parents); err != nil {
+		return err
+	}
+
+	// VerifyRandao must be after verifySigner because it needs the signer (proposer) address
+	if chain.Config().IsRandaoForkEnabled(header.Number) {
+		prevMixHash := headerMixHash(chain, parent)
+		if err := sb.VerifyRandao(chain, header, prevMixHash); err != nil {
+			return err
+		}
+	} else if header.RandomReveal != nil || header.MixHash != nil {
+		return errUnexpectedRandao
+	}
+
+	// Verify the existence / non-existence of osaka-specific header fields
+	osaka := chain.Config().IsOsakaForkEnabled(header.Number)
+	if !osaka {
+		switch {
+		case header.ExcessBlobGas != nil:
+			return errUnexpectedExcessBlobGasBeforeOsaka
+		case header.BlobGasUsed != nil:
+			return errUnexpectedBlobGasUsedBeforeOsaka
+		}
+	} else {
+		if err := eip4844.VerifyEIP4844Header(chain.Config(), parent, header); err != nil {
+			return err
+		}
+	}
+
+	return sb.verifyCommittedSeals(chain, header, parents)
+}
+
+// verifySigner checks whether the signer is in parent's validator set
+func (sb *backend) verifySigner(chain consensus.ChainReader, header *types.Header, parents []*types.Header) error {
+	// Verifying the genesis block is not supported
+	number := header.Number.Uint64()
+	if number == 0 {
+		return errUnknownBlock
+	}
+
+	// Retrieve the snapshot needed to verify this header and cache it
+	valSet, err := sb.GetValidatorSet(number)
+	if err != nil {
+		return err
+	}
+
+	// resolve the authorization key and check against signers
+	signer, err := ecrecover(header)
+	if err != nil {
+		return err
+	}
+
+	// Signer should be in the validator set of previous block's extraData.
+	if !valSet.Qualified().Contains(signer) {
+		return errUnauthorized
+	}
+	return nil
+}
+
+// verifyCommittedSeals checks whether every committed seal is signed by one of the parent's validators
+func (sb *backend) verifyCommittedSeals(chain consensus.ChainReader, header *types.Header, parents []*types.Header) error {
+	number := header.Number.Uint64()
+	// We don't need to verify committed seals in the genesis block
+	if number == 0 {
+		return nil
+	}
+
+	// Retrieve the snapshot needed to verify this header and cache it
+	valSet, err := sb.GetCommitteeStateByRound(number, uint64(header.Round()))
+	if err != nil {
+		return err
+	}
+
+	extra, err := types.ExtractIstanbulExtra(header)
+	if err != nil {
+		return err
+	}
+	// The length of Committed seals should be larger than 0
+	if len(extra.CommittedSeal) == 0 {
+		return errEmptyCommittedSeals
+	}
+
+	council := valSet.Council().Copy()
+	// Check whether the committed seals are generated by parent's validators
+	validSeal := 0
+	proposalSeal := istanbulCore.PrepareCommittedSeal(header.Hash())
+	// 1. Get committed seals from current header
+	for _, seal := range extra.CommittedSeal {
+		// 2. Get the original address by seal and parent block hash
+		addr, err := cacheSignatureAddresses(proposalSeal, seal)
+		if err != nil {
+			return errInvalidSignature
+		}
+		// Every validator can have only one seal. If more than one seals are signed by a
+		// validator, the validator cannot be found and errInvalidCommittedSeals is returned.
+		if council.Remove(addr) {
+			validSeal += 1
+		} else {
+			return errInvalidCommittedSeals
+		}
+	}
+
+	// The length of validSeal should be larger than number of faulty node + 1
+	if validSeal <= 2*valSet.F() {
+		return errInvalidCommittedSeals
+	}
+
+	return nil
+}
